update to 1.033.C
MSG_ANNOUNCE_CARD_FILTER is not implemented

+++ Game/GameBehavior.cs
@@ -665,7 +665,7 @@ namespace WindBot.Game
         {
             packet.ReadByte(); // player
             int type = packet.ReadInt16();
-            int quantity = packet.ReadByte();
+            int quantity = packet.ReadInt16();
 
             IList<ClientCard> cards = new List<ClientCard>();
             IList<int> counters = new List<int>();
@@ -676,7 +676,7 @@ namespace WindBot.Game
                 int player = GetLocalPlayer(packet.ReadByte());
                 CardLocation loc = (CardLocation) packet.ReadByte();
                 int seq = packet.ReadByte();
-                int num = packet.ReadByte();
+                int num = packet.ReadInt16();
                 cards.Add(_duel.GetCard(player, loc, seq));
                 counters.Add(num);
             }
